move version utilites to lib/version.rs
nh_lib::version
